pyproject.toml
[project]
name = "issueclaw"
version = "0.1.0"
description = "Issues as Code. Bidirectional sync between Linear and Git."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Aviad Rozenhek" }
]
keywords = [
    "linear",
    "issues",
    "cli",
    "project-management",
    "sync",
    "issues-as-code",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[project.urls]
Homepage = "https://github.com/aviadr1/issueclaw"
Repository = "https://github.com/aviadr1/issueclaw"
Issues = "https://github.com/aviadr1/issueclaw/issues"

[dependency-groups]
dev = [
    "basedpyright>=1.31.3",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.11",
]

[project.scripts]
issueclaw = "issueclaw.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (integration tests with real API calls)",
    "integration: marks tests as integration tests",
]
addopts = "--strict-markers -n auto"

[tool.basedpyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
